api-gateway/requirements.txt
fastapi
uvicorn[standard]
fastapi-users[sqlalchemy]
asyncpg
python-dotenv
alembic
pydantic-settings
psycopg2-binary